refactor: remove deterministic summarizer + legacy telegram-direct path (Path B only)
Collapses notification summary/distribution to a single path. Studio /api/notify
(Path B) summarizes AND sends; the gateway only forwards raw execution data via
ContextMemorySummarizer.BuildNotifyPayload. The local deterministic summarizer was
reachable solely through the legacy gateway-side Telegram branch in vm_runner_rest.go
('moves to /api/notify in Phase 3'), which is now removed — telegram nodes route through
/api/notify like email. SendGrid is fully gone.

Removed (~10.5k lines):
- summarizer_deterministic.go, summarizer_format_{email,telegram,discord,plaintext}.go
- ComposeSummary/ComposeSummarySmart, FormatForMessageChannels, SendGridDynamicData,
  the Summary type, ContextMemorySummarizer.Summarize (the /api/summarize HTTP call),
  buildStatusHtml + summaryForClient, the legacy 'telegram' provider branch
- the format/deterministic test files + the SendGrid integration test + the two
  obsolete vm_runner_rest_test summarize-injection tests

Kept: BuildNotifyPayload + buildRequest + the request-side types + ExtractStep helpers
(node_utils.go) + AnalyzeExecutionResult/mapExecutionStatusToAPIString (vm.go). Moved
formatValueConcise (used by vm_runner_branch.go) into summarizer.go.

Behavior change: pre-migration workflows whose Telegram node POSTs directly to
api.telegram.org lose summary injection until re-saved (they repoint to /api/notify).

Verified: go build ./... ok; go vet ok; full taskengine suite passes except the two
pre-existing failures (TestSimulateTask_StopLossWorkflow_Sepolia, TestBalanceNode_MissingAPIKey);
make storage-check clean.

Note: globalFeeRates/globalPriceService are now set-but-unread (only the deleted
summarizer read them); left wired to avoid touching engine fee-estimation startup —
follow-up cleanup.
