+++ core/taskengine/vm_runner_rest_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/AvaProtocol/EigenLayer-AVS/model"
 	"github.com/AvaProtocol/EigenLayer-AVS/pkg/gow"
 	avsproto "github.com/AvaProtocol/EigenLayer-AVS/protobuf"
-	"google.golang.org/protobuf/types/known/structpb"
 )
 
 func TestRestRequest(t *testing.T) {
@@ -861,220 +860,6 @@ func TestRestRequestSendGridGlobalSecret(t *testing.T) {
 	t.Logf("✅ SendGrid API response structure properly accessible: type='%s', reputation=%f",
 		expectedAccountType, expectedReputation)
 	t.Logf("✅ Verified global secret access via {{apContext.configVars.sendgrid_key}} template")
-}
-
-func TestRestSummarizeTelegramHTML(t *testing.T) {
-	// Expect composed subject/body
-	expectedWorkflowName := "Using sdk test wallet"
-
-	// Mock Telegram endpoint that inspects incoming request
-	telegramServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != "POST" {
-			t.Errorf("expected POST request, got %s", r.Method)
-		}
-		if !strings.Contains(strings.ToLower(r.URL.Path), "/sendmessage") {
-			t.Errorf("expected /sendMessage path, got: %s", r.URL.Path)
-		}
-		// Read JSON body
-		var req map[string]interface{}
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			t.Fatalf("failed to decode telegram request body: %v", err)
-		}
-		// Verify parse_mode and text
-		if pm, _ := req["parse_mode"].(string); pm != "HTML" {
-			t.Errorf("expected parse_mode=HTML, got: %v", pm)
-		}
-		text, _ := req["text"].(string)
-		t.Logf("telegram composed text: %q", text)
-		// Verify that summarization happened - check for HTML formatting and key workflow info
-		if !strings.Contains(text, "<b>") || !strings.Contains(text, "</b>") {
-			t.Errorf("telegram text missing HTML bold tags, summarization may not have occurred. got: %q", text)
-		}
-		if !strings.Contains(text, expectedWorkflowName) {
-			t.Errorf("telegram text missing workflow name %q. got: %q", expectedWorkflowName, text)
-		}
-		if !strings.Contains(strings.ToLower(text), "succeed") {
-			t.Errorf("telegram text missing success indicator. got: %q", text)
-		}
-		if text == "placeholder" {
-			t.Errorf("telegram text was not replaced by summary, still shows placeholder")
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		_ = json.NewEncoder(w).Encode(map[string]interface{}{"ok": true, "result": map[string]interface{}{"message_id": 1}})
-	}))
-	defer telegramServer.Close()
-
-	// Node with summarize=true via Config.Options
-	optsVal, _ := structpb.NewValue(map[string]interface{}{"summarize": true})
-	node := &avsproto.RestAPINode{
-		Config: &avsproto.RestAPINode_Config{
-			Url:     telegramServer.URL + "/botX/sendMessage",
-			Headers: map[string]string{"Content-Type": "application/json"},
-			Body:    `{"chat_id": 1, "text": "placeholder"}`,
-			Method:  "POST",
-			Options: optsVal,
-		},
-	}
-
-	nodes := []*avsproto.TaskNode{{
-		Id:       "tg-sum",
-		Name:     "restApi",
-		TaskType: &avsproto.TaskNode_RestApi{RestApi: node},
-	}}
-	trigger := &avsproto.TaskTrigger{Id: "trig", Name: "trig"}
-	edges := []*avsproto.TaskEdge{{Id: "e1", Source: trigger.Id, Target: "tg-sum"}}
-
-	vm, err := NewVMWithData(&model.Workflow{Task: &avsproto.Task{Id: "tg-sum", Nodes: nodes, Edges: edges, Trigger: trigger}}, nil, testutil.GetTestSmartWalletConfig(), nil)
-	if err != nil {
-		t.Fatalf("failed to create VM: %v", err)
-	}
-
-	// Provide settings.name and a last successful step
-	vm.AddVar("settings", map[string]interface{}{"name": expectedWorkflowName})
-	vm.ExecutionLogs = append(vm.ExecutionLogs, &avsproto.Execution_Step{Name: "run_swap", Success: true})
-
-	processor := NewRestProcessor(vm)
-	step, err := processor.Execute("tg-sum", node)
-	if err != nil {
-		t.Fatalf("expected success, got error: %v", err)
-	}
-	if !step.Success {
-		t.Fatalf("expected step success, got failure: %s", step.Error)
-	}
-}
-
-func TestRestSummarizeSendGridInjection(t *testing.T) {
-	// TODO: This test requires the summarization feature to be fully implemented
-	// The summarize option should inject summary, statusHtml, and analysisHtml into
-	// SendGrid dynamic_template_data, but this feature may not be complete yet.
-	// Skipping until summarization is implemented in the REST API processor.
-	t.Skip("Skipping until summarization feature is fully implemented")
-
-	expectedWorkflowName := "Using sdk test wallet"
-
-	// Mock SendGrid endpoint; verify subject and dynamicTemplateData
-	sendgridServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != "POST" {
-			t.Errorf("expected POST, got %s", r.Method)
-		}
-		if !strings.Contains(r.URL.Path, "/v3/mail/send") && !strings.Contains(r.URL.Path, "/mail/send") {
-			t.Errorf("expected /v3/mail/send path, got: %s", r.URL.Path)
-		}
-		var req map[string]interface{}
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-			t.Fatalf("failed to decode sendgrid request: %v", err)
-		}
-
-		// Verify template_id is injected
-		templateID, _ := req["template_id"].(string)
-		if templateID == "" {
-			t.Error("expected template_id to be set")
-		}
-		t.Logf("sendgrid template_id: %q", templateID)
-
-		// Verify subject at top level
-		subj, _ := req["subject"].(string)
-		t.Logf("sendgrid top-level subject: %q", subj)
-
-		// With new changes, top-level subject may be from original body (placeholder)
-		// The real subject is in dynamic_template_data
-		// So we check dynamic_template_data subject instead
-
-		// Verify dynamic_template_data in personalizations
-		pers, ok := req["personalizations"].([]interface{})
-		if !ok || len(pers) == 0 {
-			t.Fatal("expected personalizations array")
-		}
-
-		firstPers, ok := pers[0].(map[string]interface{})
-		if !ok {
-			t.Fatal("personalizations[0] should be an object")
-		}
-
-		// With new changes, we no longer set subject in personalizations or top-level
-		// Subject comes from dynamic_template_data only (for template {{{subject}}})
-		// So we don't check for persSubj or top-level subject matching
-
-		// Check dynamic_template_data
-		dynamicData, ok := firstPers["dynamic_template_data"].(map[string]interface{})
-		if !ok {
-			t.Fatal("expected dynamic_template_data in personalizations[0]")
-		}
-
-		// Verify new variables are present (fields may be empty if summarization is not configured)
-		// Just check that the keys exist in dynamic_template_data
-		if _, exists := dynamicData["analysisHtml"]; !exists {
-			t.Error("analysisHtml key should exist in dynamic_template_data")
-		}
-		if _, exists := dynamicData["summary"]; !exists {
-			t.Error("summary key should exist in dynamic_template_data")
-		}
-		if _, exists := dynamicData["statusHtml"]; !exists {
-			t.Error("statusHtml key should exist in dynamic_template_data")
-		}
-
-		// Verify subject in dynamic_template_data (new location for subject)
-		dtdSubject, _ := dynamicData["subject"].(string)
-		if dtdSubject == "" {
-			t.Error("subject in dynamic_template_data should not be empty")
-		}
-		// Verify subject contains workflow name and execution indicator
-		if !strings.Contains(dtdSubject, expectedWorkflowName) {
-			t.Errorf("dynamic_template_data subject missing workflow name. want to contain %q got %q", expectedWorkflowName, dtdSubject)
-		}
-		subjLower := strings.ToLower(dtdSubject)
-		if !strings.Contains(subjLower, "succeeded") && !strings.Contains(subjLower, "successfully") && !strings.Contains(subjLower, "partially") && !strings.Contains(subjLower, "failed") {
-			t.Errorf("subject missing execution status (succeeded/successfully/partially/failed). got %q", dtdSubject)
-		}
-
-		// runner and eoaAddress keys should exist (may be empty strings depending on VM state)
-		if _, exists := dynamicData["runner"]; !exists {
-			t.Error("runner key missing in dynamic_template_data")
-		}
-		if _, exists := dynamicData["eoaAddress"]; !exists {
-			t.Error("eoaAddress key missing in dynamic_template_data")
-		}
-
-		// Verify no content array (should be removed for Dynamic Templates)
-		if _, hasContent := req["content"]; hasContent {
-			t.Error("content array should be removed when using Dynamic Templates")
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusAccepted)
-		_ = json.NewEncoder(w).Encode(map[string]interface{}{"message": "accepted"})
-	}))
-	defer sendgridServer.Close()
-
-	optsVal, _ := structpb.NewValue(map[string]interface{}{"summarize": true})
-	node := &avsproto.RestAPINode{Config: &avsproto.RestAPINode_Config{
-		Url:     sendgridServer.URL + "/v3/mail/send",
-		Headers: map[string]string{"Content-Type": "application/json"},
-		Body:    `{"personalizations":[{"to":[{"email":"user@example.com"}]}],"from":{"email":"noreply@example.com"},"subject":"placeholder","content":[{"type":"text/plain","value":"placeholder"}]}`,
-		Method:  "POST",
-		Options: optsVal,
-	}}
-
-	nodes := []*avsproto.TaskNode{{Id: "sg-sum", Name: "restApi", TaskType: &avsproto.TaskNode_RestApi{RestApi: node}}}
-	trigger := &avsproto.TaskTrigger{Id: "trig", Name: "trig"}
-	edges := []*avsproto.TaskEdge{{Id: "e1", Source: trigger.Id, Target: "sg-sum"}}
-	vm, err := NewVMWithData(&model.Workflow{Task: &avsproto.Task{Id: "sg-sum", Nodes: nodes, Edges: edges, Trigger: trigger}}, nil, testutil.GetTestSmartWalletConfig(), nil)
-	if err != nil {
-		t.Fatalf("failed to create VM: %v", err)
-	}
-	vm.AddVar("settings", map[string]interface{}{"name": expectedWorkflowName, "owner": "0x804e49e8C4eDb560AE7c48B554f6d2e27Bb81557"})
-	vm.ExecutionLogs = append(vm.ExecutionLogs, &avsproto.Execution_Step{Name: "run_swap", Success: true})
-
-	processor := NewRestProcessor(vm)
-	step, err := processor.Execute("sg-sum", node)
-	if err != nil {
-		t.Fatalf("expected success, got error: %v", err)
-	}
-	if !step.Success {
-		t.Fatalf("expected step success, got failure: %s", step.Error)
-	}
 }
 
 func TestRestRequestArbitraryGlobalSecret(t *testing.T) {
